Solved with ports - run the following command to start servers parallel: npm start
