+++ my-app/package.json
@@ -18,6 +18,8 @@
     "antd": "^4.6.3",
     "express": "^4.17.1",
     "formik": "^2.1.4",
+    "nodemon": "^2.0.4",
+    "npm-run-all": "^4.1.5",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-router": "^5.2.0",
@@ -28,7 +30,9 @@
     "typescript": "^3.7.5"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "npm-run-all --parallel start-server start-client",
+    "start-client": "react-scripts start",
+    "start-server": "nodemon server/server.js",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
